Crate selection

Selection algorithms

This crate contains implementations of various algorithms for random selection of block-producer committees used by the Partner Chains Toolkit.

Modules

ariadne

Random selection out of two pools of weighted (trustless) and unweighted (permissioned) candidates controlled by a T/P ratio (D-parameter)

ariadne_v2

Random selection out of two pools of weighted (trustless) and unweighted (permissioned) candidates controlled by a T/P ratio (D-parameter), with guaranteed seat allocations

weighted_random

Simple independent weighted random selection

Type Aliases

Weight

Weight of individual candidate